+++ nemo/core/neural_factory.py
@@ -20,13 +20,11 @@ __all__ = [
     'OperationMode',
     'Optimization',
     'DeviceType',
-    'Actions',
     'NeuralModuleFactory',
     'DeploymentFormat',
 ]
 
 import random
-from abc import ABC, abstractmethod
 from enum import Enum
 from typing import List, Optional, Union
 
@@ -36,200 +34,7 @@ import nemo
 from nemo.core.callbacks import ActionCallback, EvaluatorCallback, NeMoCallback
 from nemo.core.neural_types import NmTensor
 from nemo.utils import ExpManager, logging
-from nemo.utils.app_state import AppState
 from nemo.utils.decorators import deprecated
-
-
-class TrainingState:
-    def __init__(self, action: 'Actions'):
-        """A class used to wrap the current training state of an Actions.train() function. This class holds a mapping
-        of tensor.unique_name -> it's backend tensor (eg Pytorch Tensor) or None if the tensor has been been computed
-        on the current step.
-
-        args:
-            action (Actions): The Actions object this state is associated with.
-        """
-        tensor_naming_registery = AppState().tensor_names
-        self.tensor_dict = {}.fromkeys(tensor_naming_registery.unique_names, None)
-        self._action = action
-
-    def tensor_list(self):
-        """Returns a list the unique names of all tensors.
-        """
-        return self.tensor_dict.keys()
-
-    def clear_dict(self):
-        """Clears the dictionary by setting all values to None. Used in-between training batches to clear it's state.
-        """
-        for name in self.tensor_dict:
-            self.tensor_dict[name] = None
-
-    def set_tensor(self, tensor: NmTensor, value: 'torch.Tensor'):
-        """Sets the value of tensor
-
-        args:
-            tensor (NmTensor)
-            value (torch.Tensor)
-        """
-        self.tensor_dict[tensor.unique_name] = value
-
-    def check_tensor_cached(self, unique_name: str):
-        """Checks to see the tensor value has been computed in the current step yet.
-
-        args:
-            unique_name (str): The NmTensor.unique_name that we want to check for.
-
-        returns:
-            (bool) whether the tensor with unique_name has been computed yet.
-        """
-        if self.tensor_dict[unique_name] is None:
-            return False
-        return True
-
-    def get_tensor(self, name: Union[str, NmTensor], compute: bool = True):
-        """Returns the value associated with a tensor. And optionally, computes the value of the tensor if not already
-        set.
-
-        args:
-            name (str, NmTensor): The user-defined name for a tensor or the NmTensor itself.
-            compute (bool): If True and the tensor has not already been computed, there will be an attempt to create a
-                call DAG and then do a forward pass on this call DAG to compute the tensor. If False, it will return
-                None if the tensor has not been computed yet.
-                Defaults to True.
-
-        returns:
-            (torch.tensor or None) representing the computed value of the requested name. Returns None if compute is
-            False and the tensor has not been computed yet.
-        """
-        if isinstance(name, NmTensor):
-            unique_name = name.unique_name
-        else:
-            unique_name = AppState().tensor_names[name]
-        tensor_value = self.tensor_dict[unique_name]
-        if tensor_value is None and compute:
-            nmtensor = AppState().tensor_names._nmtensor_uniname_dict[unique_name]
-            callchain = topological_sort_from_leaves([nmtensor], cached_training_state=self)
-            callchain.insert(0, ())
-            self._action.nm_graph_forward_pass(callchain, self.tensor_dict)
-            tensor_value = self.tensor_dict[unique_name]
-        return tensor_value
-
-
-def topological_sort_from_leaves(leaf_nmtensors: List[NmTensor], cached_training_state: TrainingState = None):
-    """A function that accepts a list of NmTensors that need to be computed and constructs a callchain DAG that starts
-    from a datalayerNM and can be used to compute the NmTensors.
-
-    args:
-        leaf_nmtensors (List[NmTensors]): The tensors to be computed
-        cached_training_state (TrainingState): A dictionary of already computed tensors.
-            Defaults to None meaning an empty cache.
-
-    returns:
-        top_sorted_modules: the callchain DAG
-    """
-    from nemo.backends.pytorch.nm import DataLayerNM  # TODO: Replace this with a backend agnostic data layer
-
-    def create_node(producer, producer_args):
-        if producer_args is None:
-            return tuple((producer, ()))
-        else:
-            return tuple((producer, tuple([(k, v) for k, v in producer_args.items()]),))
-
-    def is_in_degree_zero(node, processed_nodes, cached_training_state):
-        """A node has in degree of zero"""
-        if node[1] == ():
-            return True
-        for portname, nmtensor in node[1]:
-            nd = create_node(nmtensor.producer, nmtensor.producer_args)
-            if nd not in processed_nodes:
-                if cached_training_state and cached_training_state.check_tensor_cached(nmtensor.unique_name):
-                    continue
-                return False
-        return True
-
-    hooks = leaf_nmtensors if isinstance(leaf_nmtensors, list) else [leaf_nmtensors]
-
-    # ensures that no tensors are processed twice
-    processed_nmtensors = set()
-
-    indices_to_remove = []
-    # Check for duplicates in hook
-    for i, nmtensor in enumerate(hooks):
-        if nmtensor in processed_nmtensors:
-            indices_to_remove.append(i)
-        else:
-            processed_nmtensors.add(nmtensor)
-
-    for i in reversed(indices_to_remove):
-        hooks.pop(i)
-
-    _top_sorted_modules = []
-    all_nodes = {}
-
-    # extract all nodes to all_nodes set
-    hooks_lst = list(hooks)
-    while len(hooks_lst) > 0:
-        # take nmtensor from the end of the list
-        nmtensor = hooks_lst.pop()
-        producer_args = nmtensor.producer_args
-
-        node = create_node(nmtensor.producer, producer_args)
-        # Store nmtensor as an output of its producer
-        # first make sure all keys are present per output port
-        # and nm is inside all_nodes
-        if node not in all_nodes:
-            all_nodes[node] = {k: None for k in nmtensor.producer.output_ports}
-        # second, populate output port with current nmtensor
-        # where applicable
-        all_nodes[node][nmtensor.name] = nmtensor
-        processed_nmtensors.add(nmtensor)
-
-        new_tensors = set()
-        if producer_args is not None and producer_args != {}:
-            for _, new_nmtensor in producer_args.items():
-                if new_nmtensor not in processed_nmtensors:
-                    new_tensors.add(new_nmtensor)
-
-        # TODO
-        if cached_training_state:
-            for name, input_nmtensor in producer_args.items():
-                if cached_training_state.check_tensor_cached(input_nmtensor.unique_name):
-                    new_tensors.remove(input_nmtensor)
-
-        for new_nmtensor in new_tensors:
-            # put in the start of list
-            hooks_lst.insert(0, new_nmtensor)
-
-    all_node_with_output = []
-    # Iterate over all_nodes to create new nodes that include its output
-    # now all nodes have (module, input tensors, output tensors)
-    for node in all_nodes:
-        all_node_with_output.append(tuple((node[0], node[1], all_nodes[node])))
-
-    processed_nodes = []
-    while len(all_node_with_output) > 0:
-        for node in all_node_with_output.copy():
-            # if node's in_degree is zero it can be added to
-            # _top_sorted_modules
-            # this will also reduce in_degree of its children
-            if is_in_degree_zero(node, processed_nodes, cached_training_state):
-                _top_sorted_modules.append(node)
-                processed_nodes.append((node[0], node[1]))
-                all_node_with_output.remove(node)
-
-    # Create top_sorted_modules aka callchain
-    top_sorted_modules = []
-    for i, m in enumerate(_top_sorted_modules):
-        top_sorted_modules.append((m[0], dict(m[1]), m[2]))
-        # Ensure that there is only one dataset in callchain
-        if i > 0 and isinstance(m[0], DataLayerNM):
-            raise ValueError("There were more than one DataLayer NeuralModule inside your DAG.")
-
-        # TODO
-        if cached_training_state and isinstance(m[0], DataLayerNM):
-            raise ValueError("Could not compute tensor from current cached training state.")
-
-    return top_sorted_modules
 
 
 class DeploymentFormat(Enum):
@@ -273,238 +78,6 @@ class DeviceType(Enum):
     GPU = 1
     CPU = 2
     AllGpu = 3
-
-
-class Actions(ABC):
-    """Basic actions allowed on graphs of Neural Modules"""
-
-    def __init__(self, local_rank, global_rank, optimization_level=Optimization.mxprO0):
-        self._local_rank = local_rank
-        self._global_rank = global_rank
-        self._optim_level = optimization_level
-
-    @property
-    def local_rank(self):
-        """Local rank during distributed execution. None if single GPU/CPU
-
-        Returns:
-            (int) rank or worker or None if not in distributed model
-        """
-        return self._local_rank
-
-    @property
-    def global_rank(self):
-        """Global rank during distributed execution. None if single GPU/CPU
-
-        Returns:
-            (int) rank or worker or None if not in distributed model
-        """
-        return self._global_rank
-
-    @abstractmethod
-    def train(
-        self,
-        tensors_to_optimize: List[NmTensor],
-        callbacks: Optional[List[Union[ActionCallback, NeMoCallback]]],
-        lr_policy=None,
-        batches_per_step=None,
-        stop_on_nan_loss=False,
-    ):
-        """This action executes training and (optionally) evaluation.
-
-        Args:
-            tensors_to_optimize: which tensors to optimize. Typically this is
-                single loss tesnor.
-            callbacks: list of callback objects
-            lr_policy: function which should take (initial_lr, step, epoch) and
-                return learning rate
-            batches_per_step: number of mini-batches to process before one
-                optimizer step. (default: None, same as 1). Use this
-                to simulate larger batch sizes on hardware which could not fit
-                larger batch in memory otherwise. Effectively, this will make
-                "algorithmic" batch size per GPU/worker = batches_per_step*
-                batch_size
-            stop_on_nan_loss: (default: False) If set to True, the training
-                will stop if loss=nan or inf. If set to False, the training
-                will continue.
-
-        Returns:
-            None
-        """
-        pass
-
-    @abstractmethod
-    def infer(self, tensors: List[NmTensor]):
-        """This action executes inference. Nothing is optimized.
-        Args:
-          tensors: which tensors to evaluate.
-
-        Returns:
-          None
-        """
-        pass
-
-    # @abstractmethod
-    # def save_state_to(self, path: str):
-    #     """
-    #     Saves current state such as step, epoch and optimizer parameters
-    #     Args:
-    #       path:
-
-    #     Returns:
-
-    #     """
-    #     pass
-
-    # @abstractmethod
-    # def restore_state_from(self, path: str):
-    #     """
-    #     Restores state such as step, epoch and optimizer parameters
-    #     Args:
-    #       path:
-
-    #     Returns:
-
-    #     """
-    #     pass
-
-    @abstractmethod
-    def create_optimizer(self, optimizer, things_to_optimize, optimizer_params):
-        """
-        Creates an optimizer object to be use in the train() method.
-
-        Args:
-            optimizer: Specifies which optimizer to use.
-            things_to_optimize: A list of neural modules or tensors to be
-                optimized.
-            optimizer_params: Specifies the parameters of the optimizer
-
-        Returns:
-            Optimizer
-        """
-        pass
-
-    # def _perform_on_step_start(self, callbacks):
-    #     # TODO: Most of these checks can be relaxed since we enforce callbacks
-    #     # to be a list of ActionCallback objects
-    #     if callbacks is not None and isinstance(callbacks, List) and len(callbacks) > 0:
-    #         for callback in callbacks:
-    #             if isinstance(callback, ActionCallback):
-    #                 callback.on_iteration_start()
-    #             elif isinstance(callback, NeMoCallback):
-    #                 callback.on_step_start(self.state)
-    #             else:
-    #                 raise ValueError(
-    #                     "Callback was not a child of ActionCallback nor NeMoCallback and was not understood"
-    #                 )
-
-    # def _perform_on_step_end(self, callbacks):
-    #     if callbacks is not None and isinstance(callbacks, List) and len(callbacks) > 0:
-    #         for callback in callbacks:
-    #             if isinstance(callback, ActionCallback):
-    #                 callback.on_iteration_end()
-    #             elif isinstance(callback, NeMoCallback):
-    #                 callback.on_step_end(self.state)
-    #             else:
-    #                 raise ValueError(
-    #                     "Callback was not a child of ActionCallback nor NeMoCallback and was not understood"
-    #                 )
-
-    # def _perform_on_action_start(self, callbacks):
-    #     if callbacks is not None and isinstance(callbacks, List) and len(callbacks) > 0:
-    #         for callback in callbacks:
-    #             if isinstance(callback, ActionCallback):
-    #                 callback.on_action_start()
-    #             elif isinstance(callback, NeMoCallback):
-    #                 callback.on_train_start(self.state)
-    #             else:
-    #                 raise ValueError(
-    #                     "Callback was not a child of ActionCallback nor NeMoCallback and was not understood"
-    #                 )
-
-    # def _perform_on_action_end(self, callbacks):
-    #     if callbacks is not None and isinstance(callbacks, List) and len(callbacks) > 0:
-    #         for callback in callbacks:
-    #             if isinstance(callback, ActionCallback):
-    #                 callback.on_action_end()
-    #             elif isinstance(callback, NeMoCallback):
-    #                 callback.on_train_end(self.state)
-    #             else:
-    #                 raise ValueError(
-    #                     "Callback was not a child of ActionCallback nor NeMoCallback and was not understood"
-    #                 )
-
-    # def _perform_on_epoch_start(self, callbacks):
-    #     if callbacks is not None and isinstance(callbacks, List) and len(callbacks) > 0:
-    #         for callback in callbacks:
-    #             if isinstance(callback, ActionCallback):
-    #                 callback.on_epoch_start()
-    #             elif isinstance(callback, NeMoCallback):
-    #                 callback.on_epoch_start(self.state)
-    #             else:
-    #                 raise ValueError(
-    #                     "Callback was not a child of ActionCallback nor NeMoCallback and was not understood"
-    #                 )
-
-    # def _perform_on_epoch_end(self, callbacks):
-    #     if callbacks is not None and isinstance(callbacks, List) and len(callbacks) > 0:
-    #         for callback in callbacks:
-    #             if isinstance(callback, ActionCallback):
-    #                 callback.on_epoch_end()
-    #             elif isinstance(callback, NeMoCallback):
-    #                 callback.on_epoch_end(self.state)
-    #             else:
-    #                 raise ValueError(
-    #                     "Callback was not a child of ActionCallback nor NeMoCallback and was not understood"
-    #                 )
-
-    # def _perform_on_batch_start(self, callbacks):
-    #     if callbacks is not None and isinstance(callbacks, List) and len(callbacks) > 0:
-    #         for callback in callbacks:
-    #             if isinstance(callback, ActionCallback):
-    #                 continue
-    #             elif isinstance(callback, NeMoCallback):
-    #                 callback.on_epoch_start(self.state)
-    #             else:
-    #                 raise ValueError(
-    #                     "Callback was not a child of ActionCallback nor NeMoCallback and was not understood"
-    #                 )
-
-    # def _perform_on_batch_end(self, callbacks):
-    #     if callbacks is not None and isinstance(callbacks, List) and len(callbacks) > 0:
-    #         for callback in callbacks:
-    #             if isinstance(callback, ActionCallback):
-    #                 continue
-    #             elif isinstance(callback, NeMoCallback):
-    #                 callback.on_epoch_end(self.state)
-    #             else:
-    #                 raise ValueError(
-    #                     "Callback was not a child of ActionCallback nor NeMoCallback and was not understood"
-    #                 )
-
-    # def _init_callbacks(self, callbacks):
-    #     if callbacks is not None and isinstance(callbacks, List) and len(callbacks) > 0:
-    #         for callback in callbacks:
-    #             if isinstance(callback, ActionCallback):
-    #                 callback.action = self
-
-    # def _update_callbacks(
-    #     self, callbacks=None, registered_tensors=None, final_loss=None,
-    # ):
-    #     # if self.local_rank is None or self.local_rank == 0:
-    #     if callbacks is not None and isinstance(callbacks, List) and len(callbacks) > 0:
-    #         for callback in callbacks:
-    #             if isinstance(callback, ActionCallback):
-    #                 callback._registered_tensors = registered_tensors
-    #             else:  # For now, we can use the old callback function. In the future we should improve this
-    #                 self.training_state.tensor_dict["loss"] = final_loss
-
-
-def _str_to_opt_level(opt_str: str) -> Optimization:
-    number = int(opt_str[1:])
-    if number not in Optimization._value2member_map_:
-        raise ValueError(f"Unknown optimization value {opt_str}")
-    return Optimization(number)
 
 
 class NeuralModuleFactory(object):
@@ -715,21 +288,6 @@ class NeuralModuleFactory(object):
         Returns:
           NeuralModule instance
         """
-
-        # TK: "optimization_level" is not passed as parameter anymore.
-        # if params is not None and "optimization_level" in params:
-        #    if params["optimization_level"] != self._optim_level:
-        #        logging.warning(
-        #            "Module's {0} requested optimization level {1} is"
-        #            "different from the one specified by factory - {2}."
-        #            "Using: {3} for this module".format(
-        #                name, params["optimization_level"], self._optim_level, params["optimization_level"],
-        #            )
-        #        )
-        # else:
-        #    if params is None:
-        #        params = {}
-        #    params["optimization_level"] = self._optim_level
 
         if self._backend == Backend.PyTorch:
             return self.__get_pytorch_module(name=name, collection=collection, params=params, pretrained=pretrained,)
