+++ nemo/core/neural_types/__init__.py
@@ -19,4 +19,3 @@ from nemo.core.neural_types.axes import *
 from nemo.core.neural_types.comparison import *
 from nemo.core.neural_types.elements import *
 from nemo.core.neural_types.neural_type import *
-from nemo.core.neural_types.nmtensor_registry import NmTensorNameRegistry
